Code changes to support Documentum TCS required groups and required group sets. Code checks for permit_type to determine if the group is required group or required group set.
For required group, all the groups are required. An intermediary ACL is created
for each required group, and a final ACL with actual principals is created with
and-both-permit inheritance type.   

For required group set, one of the group in the group set is required group. An
intermediary ACL is created with all the groups of the group set, and a final
ACL with actual principals is created with and-both-permit inheritance type. 

Code review: https://codereview.appspot.com/8975045/