Click the first setConfig button to submit the form (b/13218191).
There are two setConfig buttons on the connector configuration form,
the first one undisplayed to handle the Enter key. I don't think it
matters, and clicking the first seems cleaner, and will work if the
extra button is ever removed.

The Documentum connector code was clicking the last input in the
document, whatever that happened to be. It was the Save Configuration
button in GSA 7.0, but in 7.2 it's an undisplayed text input outside
the connector config form.

Tested manually with 7.0 and 7.2, using Chrome, Firefox, and IE 11.

Other changes:

* Do not set the cursor style to 'wait' when submitting the form. With
  GSA 7.2 this setting is sticky across pages in the Admin Console.
  The spinning cursor would remain until the whole page was reloaded
  (clicking navigation links to go to other pages would not reset the
  cursor).

Code review: http://codereview.appspot.com/159000045/